feat: implement user and order policies with authorization checks

+++ backend/app/Http/Controllers/Api/User/UserController.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 namespace App\Http\Controllers\Api\User;
 
 use App\Enums\HttpStatus;
-use App\Enums\PermissionType;
 use App\Http\Requests\Api\User\UserFilterRequest;
 use App\Http\Requests\Api\User\UserRequest;
 use App\Http\Responses\Api\ApiResponse;
+use App\Models\User;
 use App\Services\Api\User\UserService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController {
     public function __construct(
@@ -27,12 +28,7 @@ class UserController {
     }
 
     public function show(Request $request): ApiResponse {
-        if(!$request->user()?->can(PermissionType::USERS_SHOW->value)) {
-            return new ApiResponse(
-                status: HttpStatus::UNAUTHORIZED,
-                message: __('response.unauthorized'),
-            );       
-        }
+        Gate::authorize('show', User::class);
 
         $userId = (int)$request->route('id');
 
@@ -71,15 +67,7 @@ class UserController {
     }
 
     public function delete(Request $request) {
-        if(!$request->user()?->can(PermissionType::USERS_DELETE->value)) {
-            return new ApiResponse(
-                status: HttpStatus::UNAUTHORIZED,
-                message: __('response.unauthorized'),
-            );       
-        }
-
         $userId = (int)$request->route('id');
-
         if(empty($userId)) {
             return new ApiResponse(
                 status: HttpStatus::BAD_REQUEST,
@@ -87,6 +75,7 @@ class UserController {
             );    
         }
 
+        Gate::authorize('delete', User::findOrFail($userId));
         $this->userService->delete($userId);
 
         return new ApiResponse(
